Make crate rustc wrapper itself
