--- utils/cargo-gear/rustc-wrapper/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "cargo-gear-rustc-wrapper"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-interprocess = "1.2.1"
